Hacky fix for https://github.com/alliedmodders/sourcemod/issues/666
The tier1 lib will need to be recompiled for all platforms as well.

Calls to the new functions in mods on older versions of the SDK obviously causes crashes,
and the interface isn't versioned. Use the existence of an exported function in the associated tier0
that was introduced at the same time as these as a psuedo version check.